+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 * [ENHANCEMENT]
 * [FEATURE]
 
+* [FEATURE] Add `tls.insecure-skip-verify` flag to ignore tls verification errors (PR #417) #348
+
 ## 0.12.1 / 2019-07-10
 
 ### Changes: